feat(sign): Create controlError.js
Create controlError.js to delete duplicated function about controling
Error
