+++ js/sign/controlError.js
@@ -1,0 +1,11 @@
+function appearError(el, elError, errorText) {
+  elError.classList.add("error-appear");
+  elError.textContent = errorText;
+  el.classList.add("error-border");
+}
+function disappearError(el, elError) {
+  elError.classList.remove("error-appear");
+  el.classList.remove("error-border");
+}
+
+export { appearError, disappearError };
